Move the current cache::hash_key into compiler::c
